+++ google/cloud/spanner_v1/types/type/__init__.py
